ref: Make algorithms easier composable
