+++ src/Math/BaseMath.php
@@ -1,0 +1,47 @@
+<?php
+
+namespace Famoser\Elliptic\Math;
+
+use Famoser\Elliptic\Math\Algorithm\DoubleAndAddAlways;
+use Famoser\Elliptic\Math\Calculator\CalculatorInterface;
+use Famoser\Elliptic\Math\Calculator\Primitives\PrimeField;
+use Famoser\Elliptic\Math\Calculator\UnsafePrimeCurveCalculator;
+use Famoser\Elliptic\Math\Utils\ConstSwapper;
+use Famoser\Elliptic\Primitives\Curve;
+use Famoser\Elliptic\Primitives\Point;
+
+abstract class BaseMath implements MathInterface
+{
+    public function __construct(private readonly CalculatorInterface $calculator)
+    {
+    }
+
+    public function getCurve(): Curve
+    {
+        return $this->calculator->getCurve();
+    }
+
+    public function mulG(\GMP $factor): Point
+    {
+        return $this->mul($this->calculator->getCurve()->getG(), $factor);
+    }
+
+    public function double(Point $a): Point
+    {
+        $nativeA = $this->calculator->affineToNative($a);
+        $nativeResult = $this->calculator->double($nativeA);
+
+        return $this->calculator->nativeToAffine($nativeResult);
+    }
+
+    public function add(Point $a, Point $b): Point
+    {
+        $nativeA = $this->calculator->affineToNative($a);
+        $nativeB = $this->calculator->affineToNative($b);
+        $nativeResult = $this->calculator->add($nativeA, $nativeB);
+
+        return $this->calculator->nativeToAffine($nativeResult);
+    }
+
+
+}
