+++ src/Math/Calculator/AffineAddCalculatorInterface.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Famoser\Elliptic\Math\Calculator;
+
+use Famoser\Elliptic\Primitives\Point;
+
+/**
+ * @template T
+ */
+interface AffineAddCalculatorInterface
+{
+    /**
+     * @param T $a
+     * @param Point $b
+     * @return T
+     */
+    public function addAffine(mixed $a, Point $b): mixed;
+}
