+++ src/Math/Utils/SwapperInterface.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Famoser\Elliptic\Math\Utils;
+
+interface SwapperInterface
+{
+    public function conditionalSwap(\GMP &$a, \GMP &$b, int $swapBit, int $maxBitLength): void;
+}
